restructure imports; reorder views sql
